Don't do int division in mean update computation.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/pool/trunk@1175096 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/pool2/impl/GenericKeyedObjectPool.java
@@ -1732,8 +1732,8 @@ public class GenericKeyedObjectPool<K,T> extends BaseKeyedObjectPool<K,T>
             Long time = iter.next();
             if (time != null) {
                 counter++;
-                result = result * ((counter - 1) / counter) +
-                        time.longValue()/counter;
+                result = result * ((counter - 1) / (double) counter) +
+                        time.longValue()/(double) counter;
             }
         }
         return (long) result;
